garch_hypernet_ensemble/requirements.txt
# Core ML stack
tensorflow[and-cuda]==2.15.0
statsmodels==0.14.0
catboost==1.2.2
lightgbm==4.1.0
scikit-learn==1.3.0

# Async and performance
asyncio-throttle==0.1.1
aiohttp==3.8.6
aioredis==2.0.1

# Validation and monitoring
pydantic[email]==2.5.0
pandas==2.1.0
numpy==1.24.3
shap==0.44.0
prometheus-client==0.19.0

# Infrastructure
cachetools==5.3.3
gitpython==3.1.43
joblib==1.4.2
redis==5.0.1
psutil==5.9.5

# Observability
structlog==23.2.0
loguru==0.7.2
opentelemetry-api==1.21.0

# Dev tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
mypy==1.7.1
pyyaml==6.0.1
scipy==1.11.4
